made 'bicluster' a submodule of 'cluster'
