--- examples/bicluster/README.txt
@@ -1,6 +1,0 @@
-.. _bicluster_examples:
-
-Biclustering
-------------
-
-Examples concerning the :mod:`sklearn.bicluster` package.
