+++ sklearn/cluster/bicluster/spectral.py
